+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to AoE II AI Tools will be documented in this file.
 
+## [0.1.2]
+ 1. Package Version
+    - Updated VSCode engine to 1.1.30
+ 2. Linter updates
+    - Added ERR2012 for when constants are redefined
+    - Added ERR2008 for when an arrow is missing from a rule
+    - 
+
+
 ## [0.1.1]
  1. Content Update
     - Made syntax support changes to improve the quality of this extension